+++ cmd/hidden_lake/service/pkg/config/config.go
@@ -5,11 +5,11 @@ import (
 	"sync"
 
 	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/settings"
+	"github.com/number571/go-peer/internal/logger"
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
 	"github.com/number571/go-peer/pkg/encoding"
 	"github.com/number571/go-peer/pkg/errors"
 	"github.com/number571/go-peer/pkg/filesystem"
-	"github.com/number571/go-peer/pkg/logger"
 )
 
 var (