+++ cmd/hidden_lake/service/pkg/config/types.go
@@ -1,8 +1,8 @@
 package config
 
 import (
+	"github.com/number571/go-peer/internal/logger"
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
-	"github.com/number571/go-peer/pkg/logger"
 )
 
 type IWrapper interface {