+++ CHANGELOG.md
@@ -13,6 +13,7 @@ In progress...
 - Append `examples`: routing to echo_service 
 - Create `CONNECTIONS.md`: list of connections to HLT relayers and HLS nodes
 - Update `hidden_lake/traffic`: append option 'storage'=(true|false)
+- Update `network/anonymous`: change logger -> logBuilder
 
 ### CHANGES
 
@@ -25,7 +26,7 @@ In progress...
 ### BUG FIXES
 
 - Update `README.md`: url with images -> _images
-- Update `cmd/micro-anon`: change panic error -> print error
+- Update `cmd/micro_anon`: change panic error -> print error
 - Change `hidden_lake/service`: rename CLogWarnOffResponseFromService, CLogWarnResponseFromService -> CLogInfoOffResponseFromService, CLogInfoResponseFromService
 
 ## v1.5.12